more files in ts and folder structure
