--- src/components/shared/Footer.js
@@ -1,11 +1,0 @@
-const Footer = () => {
-    return (
-        <footer className="main-footer">
-            <div className="container text-center">
-                Created by <a href="https://github.com/hidaytrahman">Hidayt Rahman</a>
-            </div>
-        </footer>
-    )
-}
-
-export default Footer;
