+++ src/components/shared/Footer.tsx
@@ -1,0 +1,11 @@
+const Footer = () => {
+  return (
+    <footer className="main-footer">
+      <div className="container text-center">
+        Created by <a href="https://github.com/hidaytrahman">Hidayt Rahman</a>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
