+++ src/components/todo/todoAdd/TodoAdd.test.js
@@ -1,16 +1,16 @@
 import { render, screen } from "@testing-library/react";
-import AddTodo from "./AddTodo";
+import TodoAdd from "./TodoAdd";
 
 describe("Todo functionality coverage", () => {
   test("input text and add button controls should exists", () => {
-    render(<AddTodo />);
+    render(<TodoAdd />);
 
     expect(screen.getByTestId("todo-text")).toBeInTheDocument();
     expect(screen.getByTestId("todo-addBtn")).toBeInTheDocument();
   });
 
   test("should have `Add` as button text", () => {
-    render(<AddTodo />);
+    render(<TodoAdd />);
     expect(screen.getByTestId("todo-addBtn").textContent).toEqual("Add");
   });
 });
